+++ src/aks-preview/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 
 Guidance
-++++++++
++++++++
 If there is no rush to release a new version, please just add a description of the modification under the *Pending* section.
 
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
@@ -52,6 +52,10 @@ Pending
 20.0.0b3
 ++++++
 * Vendor new SDK and bump API version to 2026-02-02-preview.
+
+20.0.0b3
++++++++
+* Migrate code from Azure SDK to AAZ based commands for compute operations (VM).
 
 20.0.0b2
 +++++++